refactor: Rename `…/event-detection/event-types/cloudwatchLog.js`
