+++ sdk-js/src/lib/event-detection/event-types/cloudwatch-log.js
