+++ sdk-js/src/lib/event-detection/index.js
@@ -6,7 +6,7 @@ const apiGatewayV2 = require('./event-types/api-gateway-v2');
 const customAuthorizer = require('./event-types/customAuthorizer');
 const cloudFront = require('./event-types/cloudFront');
 const cloudwatchEvent = require('./event-types/cloudwatch-event');
-const cloudwatchLogs = require('./event-types/cloudwatchLog');
+const cloudwatchLogs = require('./event-types/cloudwatch-log');
 const dynamodb = require('./event-types/dynamodb');
 const firehose = require('./event-types/firehose');
 const kinesis = require('./event-types/kinesis');
